requirements.txt
# ─── TimesFM Action Server — requirements.txt ─────────────────────────────────
# Install order matters: torch first, then timesfm[torch]

# ── Web framework ──────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ── Data ───────────────────────────────────────────────────────────────────────
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0

# ── PyTorch (CUDA 12.1 — RTX 40xx) ────────────────────────────────────────────
# Install separately with the right CUDA index:
#   uv pip install torch==2.3.1+cu121 --index-url https://download.pytorch.org/whl/cu121
# Uncomment below for CPU-only fallback:
# torch>=2.3.0

# ── TimesFM — install from source ─────────────────────────────────────────────
# git clone https://github.com/google-research/timesfm && cd timesfm
# uv pip install -e ".[torch]"
# Optional with xreg covariates support:
# uv pip install -e ".[torch,xreg]"

# ── Utilities ──────────────────────────────────────────────────────────────────
httpx>=0.27.0           # async HTTP (optional LangChain tool)
python-dotenv>=1.0.0
